refactor: placed components inside components folder, moved other files
