+++ src/app/(protected)/all-applicants/[applicantID]/page.tsx
@@ -1,6 +1,6 @@
 // app/all-applicants/[applicantID]/page.tsx
 import React from 'react';
-import ApplicantReviewDetail from '../../pages/allapplicantsmodule/ApplicantReviewDetail';
+import ApplicantReviewDetail from '../../../../components/all-applicants/ApplicantReviewDetail';
 
 export default function ApplicantDetailPage() {
   return <ApplicantReviewDetail />;
