+++ src/app/(protected)/job-listing/Result.tsx
