+++ src/app/(protected)/organization/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Organization from '../pages//organization/Organization';
+import Organization from '@/components/organization/Organization';
 
 export default function Page(): React.JSX.Element {
   return <Organization />;
