+++ src/app/(protected)/schedule/Result.tsx
