+++ src/components/all-applicants/resume-preview.tsx
